added array resets for game entities
its a little hacky but works. I check if the array exsists on entry and
create the entities. The way create function works is it initialized
everything to zero so after init i can call the create function and it
will reset the entity array. Need to do the same for the Bosses maybe.