refactor and bug fix
refactored demo state change timer back to main to be under the
rendering loop like all the other state changes. Added the poopshot to
the entity exclusion for processing as if it is off the left screen it
needs to progress untill it expires or the player won't have a way to
shoot.